Just use cached forecast data
Courtesy of how async JS works, we can’t return the value of an async
function (the callback of `forecastIo.forecast()` in this case), but we
can, thanks to the speed of the API response, just use the
cached response.

It might bite us later, and promises are almost def. a better way to go,
but for now, this works.